libzeth: move mpc code into its own subdir
