+++ libzeth/mpc/groth16/mpc_utils.tcc
@@ -2,14 +2,14 @@
 //
 // SPDX-License-Identifier: LGPL-3.0+
 
-#ifndef __ZETH_SNARKS_GROTH16_MPC_UTILS_TCC__
-#define __ZETH_SNARKS_GROTH16_MPC_UTILS_TCC__
+#ifndef __ZETH_MPC_GROTH16_MPC_UTILS_TCC__
+#define __ZETH_MPC_GROTH16_MPC_UTILS_TCC__
 
 #include "libzeth/core/evaluator_from_lagrange.hpp"
 #include "libzeth/core/multi_exp.hpp"
 #include "libzeth/core/utils.hpp"
-#include "libzeth/snarks/groth16/mpc/mpc_utils.hpp"
-#include "libzeth/snarks/groth16/mpc/phase2.hpp"
+#include "libzeth/mpc/groth16/mpc_utils.hpp"
+#include "libzeth/mpc/groth16/phase2.hpp"
 
 #include <algorithm>
 #include <exception>
@@ -258,4 +258,4 @@ srs_mpc_layer_L1<ppT> mpc_compute_linearcombination(
 
 } // namespace libzeth
 
-#endif // __ZETH_SNARKS_GROTH16_MPC_UTILS_TCC__
+#endif // __ZETH_MPC_GROTH16_MPC_UTILS_TCC__
