+++ libzeth/test/mpc_test_hash.cpp
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: LGPL-3.0+
 
 #include "libzeth/core/utils.hpp"
-#include "libzeth/snarks/groth16/mpc/hash_utils.hpp"
+#include "libzeth/mpc/groth16/hash_utils.hpp"
 
 #include <gtest/gtest.h>
