+++ libzeth/test/powersoftau_test.cpp
@@ -4,7 +4,7 @@
 
 #include "libzeth/core/evaluator_from_lagrange.hpp"
 #include "libzeth/core/utils.hpp"
-#include "libzeth/snarks/groth16/mpc/powersoftau_utils.hpp"
+#include "libzeth/mpc/groth16/powersoftau_utils.hpp"
 
 #include <boost/filesystem.hpp>
 #include <fstream>
