+++ mpc_tools/mpc_phase2/cli/mpc_create_keypair.cpp
@@ -3,9 +3,9 @@
 // SPDX-License-Identifier: LGPL-3.0+
 
 #include "libzeth/core/utils.hpp"
-#include "libzeth/snarks/groth16/mpc/mpc_utils.hpp"
-#include "libzeth/snarks/groth16/mpc/phase2.hpp"
-#include "libzeth/snarks/groth16/mpc/powersoftau_utils.hpp"
+#include "libzeth/mpc/groth16/mpc_utils.hpp"
+#include "libzeth/mpc/groth16/phase2.hpp"
+#include "libzeth/mpc/groth16/powersoftau_utils.hpp"
 #include "mpc_common.hpp"
 
 #include <vector>
