+++ mpc_tools/mpc_phase2/cli/mpc_phase2_verify_contribution.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: LGPL-3.0+
 
-#include "libzeth/snarks/groth16/mpc/phase2.hpp"
+#include "libzeth/mpc/groth16/phase2.hpp"
 #include "mpc_common.hpp"
 
 using namespace libzeth;
